refactor: extract slide validation core
