+++ tests/validation/fixtures/sample-deck/slide-01.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Pass Fixture</title>
+    <style>
+      body {
+        margin: 0;
+        width: 960px;
+        height: 540px;
+        position: relative;
+        overflow: hidden;
+        font-family: sans-serif;
+        background: #ffffff;
+      }
+
+      .card {
+        position: absolute;
+        inset: 60px 80px auto 80px;
+        padding: 24px;
+        border: 2px solid #111111;
+      }
+
+      h1,
+      p {
+        margin: 0;
+      }
+
+      p {
+        margin-top: 12px;
+        font-size: 20px;
+        line-height: 1.4;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="card">
+      <h1>Quarterly Summary</h1>
+      <p>All content stays inside the frame and no text is clipped.</p>
+    </div>
+  </body>
+</html>
