+++ tests/validation/fixtures/sample-deck/slide-02.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Overflow Fixture</title>
+    <style>
+      body {
+        margin: 0;
+        width: 960px;
+        height: 540px;
+        position: relative;
+        overflow: hidden;
+        font-family: sans-serif;
+        background: #ffffff;
+      }
+
+      .frame-breaker {
+        position: absolute;
+        top: 32px;
+        left: 930px;
+        width: 80px;
+        padding: 8px;
+        background: #ffe082;
+      }
+
+      p {
+        margin: 0;
+        font-size: 18px;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="frame-breaker">
+      <p>Outside frame</p>
+    </div>
+  </body>
+</html>
